Don't re-build distributary unprompted

+++ lobsters/mysql/src/ec2.rs
@@ -32,23 +32,25 @@ impl fmt::Display for Backend {
 }
 
 fn git_and_cargo(ssh: &mut Session, dir: &str, bin: &str) -> Result<(), failure::Error> {
-    eprintln!(" -> git reset");
-    ssh.cmd(&format!("bash -c 'git -C {} reset --hard 2>&1'", dir))
-        .map(|out| {
-            let out = out.trim_right();
-            if !out.is_empty() && !out.contains("Already up-to-date.") {
-                eprintln!("{}", out);
-            }
-        })?;
+    if dir != "distributary" {
+        eprintln!(" -> git reset");
+        ssh.cmd(&format!("bash -c 'git -C {} reset --hard 2>&1'", dir))
+            .map(|out| {
+                let out = out.trim_right();
+                if !out.is_empty() && !out.contains("Already up-to-date.") {
+                    eprintln!("{}", out);
+                }
+            })?;
 
-    eprintln!(" -> git update");
-    ssh.cmd(&format!("bash -c 'git -C {} pull 2>&1'", dir))
-        .map(|out| {
-            let out = out.trim_right();
-            if !out.is_empty() && !out.contains("Already up-to-date.") {
-                eprintln!("{}", out);
-            }
-        })?;
+        eprintln!(" -> git update");
+        ssh.cmd(&format!("bash -c 'git -C {} pull 2>&1'", dir))
+            .map(|out| {
+                let out = out.trim_right();
+                if !out.is_empty() && !out.contains("Already up-to-date.") {
+                    eprintln!("{}", out);
+                }
+            })?;
+    }
 
     eprintln!(" -> rebuild");
     ssh.cmd(&format!(
